+++ Changelog.md
@@ -5,6 +5,8 @@
 ### Unreleased
 ##### 2020-01-XX
 
+- Decimal interface has been added
+- All prices are decimal
 - Payable interface has been added
 - Shippable interface has been added
 - BC: Address interface has been extended and refactored to use return types